% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp_data.R
\docType{data}
\name{mgp154}
\alias{mgp154}
\title{Gut microbiota in the Irish Elderly}
\format{
An data.frame with genus and predictors.
}
\source{
\href{https://www.mg-rast.org/mgmain.html?mgpage=project&project=mgp154}{MG-RAST-mgp154}
}
\usage{
data(mgp154)
}
\description{
This study is based on pyrosequencing of 16S rDNA amplicons from faecal samples collected from 178 elderly Irish citizens and 13 healthy young control subjects. A subset of these samples were also subjected to shotgun sequencing using Illumina HiSeq 2000 2x91bp reads. Antibiotic treatment was an exclusion criterion.
}
\references{
Claesson, Marcus J., et al. "Gut microbiota composition correlates with diet and health in the elderly." Nature 488.7410 (2012): 178-184.
}
\keyword{datasets}
