% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp_data.R
\docType{data}
\name{mgp2592}
\alias{mgp2592}
\title{Hofmockel Soil Aggregate COB KBASE}
\format{
An data.frame with genus and predictors.
}
\source{
\href{https://www.mg-rast.org/mgmain.html?mgpage=project&project=mgp2592}{MG-RAST-mgp2592}
}
\usage{
data(mgp2592)
}
\description{
This study is to examine soil microbial community composition and structure of both bacteria and fungi at a microbially-relevant scale. The researchers isolated soil aggregates from three land management systems in central Iowa to test if the aggregate-level microbial responses are related to plant community and management practices. The clean dataset has 120 samples with 17 genus under consideration.
}
\references{
Bach, Elizabeth M., et al. "Greatest soil microbial diversity found in micro-habitats." Soil biology and Biochemistry 118 (2018): 217-226.
}
\keyword{datasets}
