% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.bglasso_out}
\alias{plot.bglasso_out}
\title{plot the graph estimated by graphical lasso with threshold method using ggraph}
\usage{
\method{plot}{bglasso_out}(x, ...)
}
\arguments{
\item{x}{The bglasso_out}

\item{...}{\itemize{
\item{\code{tol}}{: threshold for ploting default 0.01, if horseshoed, then horseshoe result is used}
}}
}
\value{
A \code{ggplot} xect
}
\description{
plot the graph estimated by graphical lasso with threshold method using ggraph
}
