% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.carlasso_out}
\alias{plot.carlasso_out}
\title{plot the chain graph estimated by CAR-LASSO with threshold or horseshoe method using ggraph}
\usage{
\method{plot}{carlasso_out}(obj, tol = 0.01)
}
\arguments{
\item{obj}{The carlasso_out object}

\item{tol}{threshold default 0.01, if horseshoed, then horseshoe result is used}
}
\value{
A \code{ggplot} object
}
\description{
plot the chain graph estimated by CAR-LASSO with threshold or horseshoe method using ggraph
}
