% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simu_AR1}
\alias{simu_AR1}
\title{Simulate a simple AR1 model}
\usage{
simu_AR1(n = 100, k = 5, rho = 0.7)
}
\arguments{
\item{n}{sample size}

\item{k}{number of responses}
}
\value{
rho partial correlation in AR1
}
\description{
Simulate a simple AR1 model
}
