% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp_data.R
\docType{data}
\name{mgp2592}
\alias{mgp2592}
\title{Hofmockel Soil Aggregate COB KBASE}
\format{
An data.frame with genus and predictors.
}
\source{
\href{https://www.mg-rast.org/mgmain.html?mgpage=project&project=mgp2592}{MG-RAST-mgp2592}
}
\usage{
data(mgp2592)
}
\description{
This study is based on pyrosequencing of 16S rDNA amplicons from faecal samples collected from 178 elderly Irish citizens and 13 healthy young control subjects. A subset of these samples were also subjected to shotgun sequencing using Illumina HiSeq 2000 2x91bp reads. Antibiotic treatment was an exclusion criterion.
}
\references{
Bach, Elizabeth M., et al. "Greatest soil microbial diversity found in micro-habitats." Soil biology and Biochemistry 118 (2018): 217-226.
}
\keyword{datasets}
