% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.carlasso_out}
\alias{plot.carlasso_out}
\title{plot the chain graph estimated by CAR-LASSO with threshold or horseshoe method using ggraph}
\usage{
\method{plot}{carlasso_out}(x, ...)
}
\arguments{
\item{x}{The carlasso_out xect}

\item{...}{\itemize{
\item{\code{tol}}{: threshold for ploting default 0.01, if horseshoed, then horseshoe result is used}
}}
}
\value{
A \code{ggplot} xect
}
\description{
plot the chain graph estimated by CAR-LASSO with threshold or horseshoe method using ggraph
}
